% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dataprovider.R
\docType{class}
\name{DataProvider}
\alias{DataProvider}
\title{DataProvider}
\value{
DataProvider object
}
\description{
An R6 class responsible for providing data to the application.

API Usage:

"'GET /':   Attribution and instructions of what URL paths and queries are available, along with a description of each.", 
"'GET /datasets/':   Metadata of each dataset, including spreadsheet *key*, *indicatorName*, *category* and *subcategory* the dataset belongs to, and the *dataprovider* and the *dataprovider_link*.", 
"'GET /datasets/categories/':   Shows a count of how many datasets belong to each category and subcategory.", 
"'GET /datasets/categorized/':   Shows which datasets belong to each category and subcategory.", 
"'GET /data/help/?key=key':   Shows which countries and years are available in the specified dataset.", 
"'GET /data/?key=key&country=country&year=year':   Gets the value for the specified country and year in the dataset.", 
"'GET /data/?key=key&country=country':   Gets the values for each available year in the dataset, for the specified country.", 
"'GET /data/?key=key&year=year':   Gets the values for each available country in the dataset, for the specified year.", 
"'GET /data/?key=key':   Downloads the CSV file of the whole dataset."
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{api_url}}{API URL for getting data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataProvider$new()}}
\item \href{#method-call_api}{\code{DataProvider$call_api()}}
\item \href{#method-describe_datasets}{\code{DataProvider$describe_datasets()}}
\item \href{#method-clone}{\code{DataProvider$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a DataProvider
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProvider$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a new `DataProvider` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-call_api"></a>}}
\if{latex}{\out{\hypertarget{method-call_api}{}}}
\subsection{Method \code{call_api()}}{
Call API
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProvider$call_api(endpoint)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint}}{API Endpoint}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a parsed R object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-describe_datasets"></a>}}
\if{latex}{\out{\hypertarget{method-describe_datasets}{}}}
\subsection{Method \code{describe_datasets()}}{
Describe available datasets
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProvider$describe_datasets()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a `tibble` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
