% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_DDFProvider.R
\docType{class}
\name{DDFProvider}
\alias{DDFProvider}
\title{DDFProvider}
\value{
DDFProvider object
}
\description{
An R6 class responsible for providing data to the application.

API Usage:

"'GET /':   Attribution and instructions of what URL paths and queries are available, along with a description of each.", 
"'GET /datasets/':   Metadata of each dataset, including spreadsheet *key*, *indicatorName*, *category* and *subcategory* the dataset belongs to, and the *dataprovider* and the *dataprovider_link*.", 
"'GET /datasets/categories/':   Shows a count of how many datasets belong to each category and subcategory.", 
"'GET /datasets/categorized/':   Shows which datasets belong to each category and subcategory.", 
"'GET /data/help/?key=key':   Shows which countries and years are available in the specified dataset.", 
"'GET /data/?key=key&country=country&year=year':   Gets the value for the specified country and year in the dataset.", 
"'GET /data/?key=key&country=country':   Gets the values for each available year in the dataset, for the specified country.", 
"'GET /data/?key=key&year=year':   Gets the values for each available country in the dataset, for the specified year.", 
"'GET /data/?key=key':   Downloads the CSV file of the whole dataset."
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{repos}}{List of repositories in GitHub Organization}

\item{\code{org}}{GitHub Organization}

\item{\code{org_url}}{Organization URL}

\item{\code{api_url}}{API URL}

\item{\code{data}}{Data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DDFProvider$new()}}
\item \href{#method-get_all_repos}{\code{DDFProvider$get_all_repos()}}
\item \href{#method-get_repo}{\code{DDFProvider$get_repo()}}
\item \href{#method-get_ddf}{\code{DDFProvider$get_ddf()}}
\item \href{#method-log}{\code{DDFProvider$log()}}
\item \href{#method-clone}{\code{DDFProvider$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
initalize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDFProvider$new(org = "open-numbers", log_level = INFO, debug = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org}}{GitHub Organization}

\item{\code{log_level}}{Log Level, see logger::log_level()}

\item{\code{debug}}{Enable debugging mode}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
DDFProvider object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_all_repos"></a>}}
\if{latex}{\out{\hypertarget{method-get_all_repos}{}}}
\subsection{Method \code{get_all_repos()}}{
get_all_repos
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDFProvider$get_all_repos(repos = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{repos}}{List of repositories to get}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.table

get_repo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_repo"></a>}}
\if{latex}{\out{\hypertarget{method-get_repo}{}}}
\subsection{Method \code{get_repo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDFProvider$get_repo(repo)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{repo}}{Name of repository to get}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.table

read_ddf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ddf"></a>}}
\if{latex}{\out{\hypertarget{method-get_ddf}{}}}
\subsection{Method \code{get_ddf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDFProvider$get_ddf(
  repo,
  resource = c("datapoints", "concepts", "entities"),
  prefetch = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{repo}}{Name of repository to read DDF files from}

\item{\code{resource}}{DDF-resource to download, default to "datapoints"}

\item{\code{prefetch}}{Prefetch all DDF-resources as a column of data.tables}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.table

log
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log"></a>}}
\if{latex}{\out{\hypertarget{method-log}{}}}
\subsection{Method \code{log()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDFProvider$log(msg = "Hello!", log_level = INFO)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{Message to log}

\item{\code{log_level}}{Log Level}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDFProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
